moved observable queue to subfolder
